+++ types/event.ts
@@ -1,0 +1,6 @@
+export enum AppEventNames {
+  /** 测试 */
+  TEST = 'TEST',
+  /** Hello */
+  HELLO = 'HELLO'
+}